+++ CHANGELOG.md
@@ -4,10 +4,9 @@ All notable changes to this project will be documented in this file. See [commit
 
 ## [0.3.4](https://github.com/Wildhoney/Schematik/compare/v0.3.3...v0.3.4) (2025-10-08)
 
-
 ### Features
 
-* removed form HTML element to make it React Native compat ([aa06792](https://github.com/Wildhoney/Schematik/commit/aa06792d7ef2b1cbec710c655b7f4b6387e3bfa9))
+- removed form HTML element to make it React Native compat ([aa06792](https://github.com/Wildhoney/Schematik/commit/aa06792d7ef2b1cbec710c655b7f4b6387e3bfa9))
 
 ## [0.3.3](https://github.com/Wildhoney/Schematik/compare/v0.3.2...v0.3.3) (2025-10-08)